stress/requirements.txt
# for locust under python 2.7 use this one:
locustio
# for locust under Python 3 use the following one (segfaults on my MacOS)
#git+https://github.com/locustio/locust.git
